+++ src/Controllers/Private/TenantUsers.php
@@ -25,6 +25,7 @@ use Bayfront\BonesService\Rbac\Models\TenantRolePermissionsModel;
 use Bayfront\BonesService\Rbac\Models\TenantsModel;
 use Bayfront\BonesService\Rbac\Models\TenantUserRolesModel;
 use Bayfront\BonesService\Rbac\Models\TenantUsersModel;
+use Bayfront\BonesService\Rbac\Models\UsersModel;
 
 class TenantUsers extends PrivateApiController implements CrudControllerInterface
 {
@@ -193,6 +194,30 @@ class TenantUsers extends PrivateApiController implements CrudControllerInterfac
     }
 
     /**
+     * @return void
+     * @throws ApiServiceException
+     * @throws BadRequestException
+     */
+    private function returnNoPermissions(): void
+    {
+
+        $tenantPermissionsModel = new TenantPermissionsModel($this->rbacService);
+
+        $query_filter = [
+            [
+                'id' => [
+                    'in' => ''
+                ]
+            ]
+        ];
+
+        $collection = $this->listResources($tenantPermissionsModel, $query_filter);
+
+        $this->respond(200, TenantPermissionCollection::create($collection['list'], $collection['config']));
+
+    }
+
+    /**
      * List tenant user permissions.
      *
      * @param array $params
@@ -228,12 +253,14 @@ class TenantUsers extends PrivateApiController implements CrudControllerInterfac
             throw new ApiServiceException($e->getMessage());
         }
 
+        $usersModel = new UsersModel($this->rbacService);
         $tenantsModel = new TenantsModel($this->rbacService);
 
         try {
 
             $tenant_user = $this->tenantUsersModel->read($params['id']); // Resource
-            $tenant_owner = $tenantsModel->getOwnerId($params['tenant']); // String (user id)
+            $user = $usersModel->read(Arr::get($tenant_user, 'user', ''));
+            $tenant = $tenantsModel->read($params['tenant']); // Resource
 
         } catch (DoesNotExistException) {
             throw new NotFoundException();
@@ -241,7 +268,25 @@ class TenantUsers extends PrivateApiController implements CrudControllerInterfac
             throw new ApiServiceException($e->getMessage());
         }
 
-        if (Arr::get($tenant_user, 'user') == $tenant_owner) { // If tenant user owns tenant
+        // Check user is enabled
+
+        if (Arr::get($user, 'enabled') !== true) {
+
+            $this->returnNoPermissions();
+            return;
+
+        }
+
+        // Check tenant is enabled
+
+        if (Arr::get($tenant, 'enabled') !== true && Arr::get($user, 'admin') !== true) {
+
+            $this->returnNoPermissions();
+            return;
+
+        }
+
+        if (Arr::get($user, 'admin') === true || Arr::get($tenant_user, 'user') == Arr::get($tenant, 'owner', '')) { // If user is admin or tenant user owns tenant
 
             // List all tenant permissions